backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
cryptography==42.0.7
garminconnect==0.2.40
anthropic==0.28.0
python-telegram-bot==21.1.1
apscheduler==3.10.4
httpx==0.27.0
python-dotenv==1.0.1
